controllers/version/metadata: Move `get_version_dependencies()` to dedicated module
